PC Engine/TurboGrafx-16 controller support
These changes add support for 2 button PC Engine/Turbografx-16 controllers, the turbo fire feature of later controllers is also supported. The Avenue6 pad is not supported, mainly because I do not have one.

Though the PC Engine/TurboGrafx-16 uses an 8 pin DIN plug instead of the DB9 featured in this drivers name, adding the support here still makes sense. The PC Engine/TurboGrafx-16 controllers use the same 74HC157 multiplexer as the Mega Drive/Genesis and function in a similar manner. The 8 pin DIN plug/socket is a standardized part and should not be too difficult to find for those wishing to build an adapter.